fix: pydantic schemas move (#228)
