+++ graphs/docs/modules/schemas.rst
@@ -1,0 +1,23 @@
+.. _schemas:
+
+#########
+ Schemas
+#########
+
+This module defines pydantic schemas, which are used to validate the
+configuration.
+
+.. automodule:: anemoi.graphs.schemas.base_graph
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
+
+.. automodule:: anemoi.graphs.schemas.node_schemas
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
+
+.. automodule:: anemoi.graphs.schemas.edge_schemas
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
