+++ graphs/src/anemoi/graphs/schemas/edge_schemas.py
@@ -18,7 +18,7 @@ from pydantic import Field
 from pydantic import PositiveFloat
 from pydantic import PositiveInt
 
-from anemoi.training.schemas.utils import BaseModel
+from anemoi.utils.schemas import BaseModel
 
 
 class KNNEdgeSchema(BaseModel):
