+++ graphs/src/anemoi/graphs/schemas/post_processors.py
@@ -15,8 +15,9 @@ from typing import Annotated
 from typing import Literal
 from typing import Union
 
-from pydantic import BaseModel
 from pydantic import Field
+
+from anemoi.utils.schemas import BaseModel
 
 LOGGER = logging.getLogger(__name__)
