+++ models/docs/modules/schemas.rst
@@ -1,0 +1,25 @@
+.. _schemas:
+
+#########
+ Schemas
+#########
+
+.. automodule:: anemoi.models.schemas.models
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
+
+.. automodule:: anemoi.models.schemas.processor
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
+
+.. automodule:: anemoi.models.schemas.encoder
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
+
+.. automodule:: anemoi.models.schemas.decoder
+   :members:
+   :no-undoc-members:
+   :show-inheritance:
