+++ models/src/anemoi/models/schemas/encoder.py
