+++ models/src/anemoi/models/schemas/processor.py
